Dividiendo la logica del controlador. Se creó la clase PatientRestController que responderá a las peticiones de administración de pacientes.
