+++ src/main/java/com/gabrego/usersapi/controller/UserRestController.java
@@ -16,11 +16,7 @@ import java.util.List;
 public class UserRestController {
 
     @Autowired
-    private PatientService patientService;
-
-    @Autowired
     private UserService userService;
-
 
     @GetMapping("/users")
     public List<User> findAll() {
@@ -85,57 +81,5 @@ public class UserRestController {
         return "Deleted user id - "+userId;
     }
 
-    @GetMapping("/users/patients")
-    public List<Patient> findAllPatient() {
-        return patientService.findAll();
-    }
-
-    @GetMapping("/users/patients/{userId}")
-    public Patient getPatient(@PathVariable int userId){
-        Patient patient = patientService.findById(userId);
-
-        if(patient == null) {
-            throw new RuntimeException("User id not found -"+userId);
-        }
-        //retornará al usuario con id pasado en la url
-        System.out.println(patient);
-        return patient;
-    }
-
-    @PostMapping("/users/patients")
-    public Patient addPatient(@RequestBody Patient patient) {
-        patient.setId(0);
-        User user = userService.findById(patient.getUser_id().getId());
-        patient.setUser_id(user);
-        //Este metodo guardará al usuario enviado
-        patientService.save(patient);
-
-        return patient;
-
-    }
-
-    @DeleteMapping("users/patients/{patientId}")
-    public int detetePatient(@PathVariable int patientId) {
-
-        Patient patient = patientService.findById(patientId);
-
-        if(patient == null) {
-            throw new RuntimeException("User id not found -"+patientId);
-        }
-
-        return patientService.deleteById(patientId);
-    }
-
-    @DeleteMapping("users/patients/userid/{userId}")
-    public int detetePatientByUserId(@PathVariable int userId) {
-
-        Patient patient = patientService.findByUserId(userId);
-
-        if(patient == null) {
-            throw new RuntimeException("User id not found -"+userId);
-        }
-
-        return patientService.deleteByUserId(userId);
-    }
     
 }
