+++ src/main/resources/application.properties
@@ -5,9 +5,6 @@ spring.datasource.url = jdbc:mysql://localhost:3306/usersapi?useUnicode=true&use
 spring.datasource.username=root
 spring.datasource.password=
 
-#spring.datasource.url = jdbc:mysql://remotemysql.com:3306/6neQfKO179?useUnicode=true&useJDBCCompliantTimezoneShift=true&useLegacyDatetimeCode=false&serverTimezone=UTC
-#spring.datasource.username=6neQfKO179
-#spring.datasource.password=prMHgbse86
 spring.jpa.hibernate.ddl-auto=none
 logging.level.org.springframework.web=DEBUG
